Refactor cart and wishlist models to use common item model and enhance type definitions
