+++ src/models/product-category-collections/item.model.ts
@@ -1,50 +1,5 @@
 import BaseApiResponse from "../common/base-api-response.model";
-import MediaResponse from "../common/media-response.model";
-import { CategoryItem } from "./category.model";
-import { CollectionItem } from "./collection.model";
-
-export interface StoneDetails {
-  _id: string;
-  type: string;
-  weight: string;
-  amount: number | null;
-  description: string;
-}
-
-export interface MetalRateDetails {
-  _id: string;
-  metalDetails: string;
-  rate: number;
-}
-
-export interface ItemModel {
-  _id: string;
-  itemName: string;
-  itemDescription?: string;
-  itemCategory?: CategoryItem;
-  itemMedia?: MediaResponse[];
-  thumbnail: MediaResponse;
-  goldPurity?: string;
-  itemSpecification?: string;
-  height?: string;
-  width?: string;
-  makingCharge?: number;
-  itemSKU?: string;
-  slug?: string;
-  gender?: string;
-  stoneDetails?: StoneDetails[];
-  collections?: CollectionItem[];
-  metalType?: string;
-  itemWeight?: number;
-  withGstPrice?: number;
-  withoutGstPrice?: number;
-  finalPrice?: number;
-  metalRateDetails?: MetalRateDetails;
-  hoverImage: MediaResponse;
-  grossWeight?: number;
-  isFeatured: boolean;
-  isAddedToCart?: boolean;
-}
+import { ItemModel } from "../common/item.model";
 
 export default interface ItemApiResponse extends BaseApiResponse {
   status: boolean;
